Add Queue for Sounds
When message is requested, checks if a sound is playing. If no sound playing, play the sound immediately. If there is a sound playing, add it to the queue.
At the end of the sound, check the queue. If empty, do nothing. If there's a sound in the queue, play it.
Stopping active audio clears the queue.